Started work on room object parsing
Set up some of the foundations for allowing for objects to be parsed for rooms.
Added a new TINTED_TEXTURED type for visual attribute, takes both a colour and texture and tints the texture that colour.
Fixed a bug in my VisualKind 'switching' methods that would have meant changes would not be rendered if loadShape() had already been called.
Added a new method to GObject that allows for the recursive colouring of a GObject's visualAttributes and its children's visualAttributes